* All parsed Strings are now instances of the new NullTerminatedString class, and all API methods that may have referenced a parsed string now use the CharSequence interface, which is implemented by both String and NullTerminatedString. This will not cause problems for most code, but if you're doing an equals() comparison between a String an a string value that was parsed from a packet, it will fail because Strings and NullTerminatedStrings can be equal. The best approach is to call .toString() on NullTerminatedString and compare that against your String. This was done to improve packet read/write symmetry. * Support for new packets: DockedPacket, ExplosionPacket, KlaxonPacket, CloakDecloakPacket, ActivateUpgradePacket. * PlayerShipDamagePacket now exposes the shipIndex and duration properties. * ObjectUpdatePacket supports mixed ObjectType payloads, as used by some custom software. Stock Artemis clients appear to handle this fine. However, if you wish to ensure that you generate packets how the Artemis server does, ObjectUpdatePacket now has a static method called segregate() which accepts a Collection of ArtemisObjects and generates a List of corresponding ObjectUpdatePackets where each has payloads with homogeneous ObjectTypes. ArtemisPlayer has a split() method that will break it into several objects, with the appropriate properties set for each ObjectType. It also has multiple update methods to only update the properties for a particular ObjectType. * Util now has caratToNewline() and newlineToCarat() methods to perform the newline escapes that Artemis does. However, you shouldn't typically need to call this as the API should perform this conversion for you automagically. * Fixed incorrect parsing due to the v2.4.0 valueInt subtype shuffle. * Fixed some places where a float color was incorrectly being parsed as an int. * Fixed incorrect nebula parsing. * Internal changes to ArtemisPlayer to store upgrade and system statuses in separate objects. * The AllShipSettingsPacket.Ship class was extracted to its own top-level class, and is now used by the SetShipSettingsPacket. * Most to "fighters" were changed to "single-seat craft," since a fighter is only one kind of single-seat craft. * A new example package contains some sample client and proxy code. The example code in README.md was replaced with a mention of the new package. * ToggleShieldsPacket was renamed to SetShieldsPacket and its functionality expanded to handle raising and lower shields as well as toggling them. * Torpedoes now have their own object class (ArtemisTorpedo) instead of using ArtemisGenericObject. * BitFields now correctly compute their widths when the number of bits is divisible by 8. (Artemis allocates an extra byte in that scenario.) Packets with "fake" extra properties to deal with the extra byte were updated accordingly. * ValueIntPacket no longer needs the subtype passed into the constructor when passing in a PacketReader; it will just read the subtype from the stream. * Classes which don't conform to the ValueIntPacket payload structure were refactored to no longer descend from that class. * PacketWriter now requires you to pass in the ObjectType of an object to ensure that the correct packet is written (since ArtemisPlayer can be of four different types). * IAN now confirms that objects which have methods with the Listener annotation are public. * Target selection packets have been renamed for consistency and clarity (e.g. CaptainSelectPacket -> CaptainTargetPacket). * Removed PacketWriter.writeObjByte(), which was no longer used. * Some tweaks were made so that JavaDoc generation can succeed, and pom.xml was modified to be able to generate the JavaDoc. * README.md was reorganized to simplify individual sections.

+++ src/main/java/com/walkertribe/ian/world/Artemis.java
@@ -16,7 +16,7 @@ public final class Artemis {
 	 * The default amount of coolant the player ship has at start. The amount of
 	 * coolant may change due to side mission rewards or custom scripting.
 	 */
-    public static final int DEFAULT_COOLANT = 8;
+    public static final byte DEFAULT_COOLANT = 8;
 
     /**
      * The maximum amount of coolant that can be applied to any one system.
